Fix race condition that could happen upon error
If there was an error eg. parsing the given revision, it could lead to
the Receiver being dropped (due to run()'s early return) before the
Sender could send the error (from the thread).

This, in turn, would lead to a panic message displaying from the thread,
which isn't very nice UX.

With this change, we now ignore all sending errors: anyway, if there is
a sending error it means that the receiver has been dropped early, and
thus that an error was already printed, so it's useless to overwhelm the
user.